refactor: separate QuizCustomizer subcomponents into multiple files
